New folder/requirements.txt
llama-index
llama-index-vector-stores-chroma
llama-index-embeddings-huggingface
llama-parse
chromadb
python-dotenv
